requirement.txt
APScheduler==3.6.3
async-timeout==4.0.2
cachelib==0.10.2
cachetools==4.1.0
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==3.1.0
dominate==2.3.5
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.2
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
python-dotenv==0.10.2
pytz==2020.1
redis==4.5.1
requests==2.28.2
rsa==4.0
six==1.16.0
spotipy==2.22.1
SQLAlchemy==1.3.17
tzlocal==2.1
urllib3==1.26.14
visitor==0.1.3
Werkzeug==2.2.3
WTForms==2.2.1
zipp==3.15.0
